Support vs Dev categorisation via intended capabilities.
Expands and formalises the previous notion of defining development "augentation" in the context of what the pipeline was built for.

Added an example categorisation of pipeline tasks including distictions between support and development via the intendend capabilities the pipeline was built for. This is not an exact definition used from any VFX house, but it is heavily inspired from experience in primarily large scale, but also small scale VFX Pipelines.